Move API adress to axiosInstance and friendsServices
