+++ src/actions/axiosInstance.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 const axiosInstance = axios.create({
-  baseURL: '',
+  baseURL: 'http://localhost:3005/',
   headers: {
     'Content-Type': 'application/json',
   },
